refactorización y posteo
