+++ src/components/places.js
@@ -1,4 +1,6 @@
-export const firstView = `
+import data from '../data/mxchilazo.js';
+
+export const placesView = `
 <div id="xView" class="container">
 
   <div id="header">
@@ -53,3 +55,29 @@ export const firstView = `
 
 </div>
 `;
+
+// placeCards
+export const places = data.place;
+
+export const placeCard = (place) => {
+  const component = `
+    <div id="one-container" class="one-container">
+      <img id="placeImg" src="${place.img}"class="img-container">
+      <div id="rectangle">
+        <h3 id="namePlace">${place.name}</h3>
+        <h4 id="delegation">${place.delegation}</h4>
+        <img id="likeIcon" src="img/like.png" alt="">
+      </div>
+    </div>
+    `;
+  return component;
+};
+
+// funcion es de first Utils
+export const setCards = (places, placeCard) => {
+  let html = '';
+  places.forEach(place => {
+    html += placeCard(place);
+  });
+  return html;
+};
