+++ src/components/retro.js
@@ -67,3 +67,27 @@ export const retroView = `
 
     </div>
 `;
+// const postCont = document.querySelector('#postContainer');
+
+// subir info a firestore
+export const newReview = async (buildReview) => {
+  //   document.querySelector('#postIt').addEventListener('click', (e) => {
+  //     e.preventDefault();
+  const name = document.querySelector('#name').value;
+  const post = document.querySelector('#review').value;
+  console.log(name, post);
+  // comenzar firebase registra posteos en fs
+  await buildReview(name, post);
+  //   });
+};
+
+// exportar de firebase para el parametro doc
+
+// export const reviewCard = (doc) => {
+//     const component = `
+//         <div id="card" class="card">
+//           <h3>${doc.data().name}</h3>
+//         </div>
+//         `;
+//     return component;
+//   };
