+++ src/components/signup.js
@@ -1,4 +1,8 @@
-export const signUp = `
+// import { onNavigate } from '../routes.js';
+
+// const rootDiv = document.getElementById('root');
+
+export const signUpView = `
 <div id="signUpViewContainer" class="container">
     <img id="logo" src="img/logo_mochilazo.png" alt="MXCHILAZO">
     <h2>CREAR CUENTA</h2>
@@ -11,3 +15,10 @@ export const signUp = `
     <input type="image" id="returnArrow" class="returnArrow" src="img/Vector.png">
   </div>
 `;
+
+export const signupFunc = (newUserAccount, onNavigate, rootDiv, lugares) => {
+  const email = document.querySelector('#signupEmail').value;
+  const password = document.querySelector('#signupPassword').value;
+  // comenzar firestore registra nuevos usuarios con auth
+  newUserAccount(email, password, onNavigate, rootDiv, lugares);
+};
