--- src/utils/firstviewUtils.js
@@ -1,7 +1,0 @@
-export const setCards = (places, placeCard) => {
-  let html = '';
-  places.forEach(place => {
-    html += placeCard(place);
-  });
-  return html;
-};
