--- src/views/postView.js
@@ -1,3 +1,0 @@
-export const posts = `
-
-`;
